requirements.txt
facebook-sdk==2.0.0
requests==2.12.3
